+++ test/core-test/quantum_objects_evo.jl
@@ -74,6 +74,14 @@
         @test (a2 + 2).data == a2.data + 2 * I
         @test a2 * 2 == 2 * a2
 
+        zero_like = zero(a2)
+        iden_like = one(a3)
+        zero_array = NullOperator(100)
+        iden_array = IdentityOperator(100)
+        @test zero_like == QobjEvo(zero_array, type = a2.type, dims = a2.dims)
+        @test typeof(zero_like.data) == typeof(zero_array)
+        @test iden_like == QobjEvo(iden_array, type = a3.type, dims = a3.dims)
+        @test typeof(iden_like.data) == typeof(iden_array)
         @test trans(trans(a2)) == a2
         @test trans(a2).data == transpose(a2.data)
         # @test adjoint(a2) ≈ trans(conj(a2)) # Currently doesn't work
@@ -119,15 +127,23 @@
               "Quantum Object Evo.:   type=SuperOperator   dims=$L_dims   size=$L_size   ishermitian=$L_isherm   isconstant=$L_isconst\n$datastring"
     end
 
-    @testset "Type Inference (QuantumObject)" begin
+    @testset "Type Inference (QobjEvo)" begin
+        N = 4
         for T in [ComplexF32, ComplexF64]
-            N = 4
             a = MatrixOperator(rand(T, N, N))
             @inferred QobjEvo(a)
             for type in [Operator, SuperOperator]
                 @inferred QobjEvo(a, type = type)
             end
         end
+
+        a = destroy(N)
+        coef1(p, t) = exp(-t)
+        coef2(p::Vector, t) = sin(p[1] * t)
+        coef3(p::NamedTuple, t) = cos(p.ω * t)
+        @inferred QobjEvo(a, coef1)
+        @inferred QobjEvo((a', coef2))
+        @inferred QobjEvo((a' * a, (a, coef1), (a', coef2), (a + a', coef3)))
 
         @testset "Math Operation" begin
             a = QobjEvo(destroy(20))
@@ -182,6 +198,7 @@
         @test isconstant(H_td) == false
         @test isconstant(QobjEvo(a)) == true
         @test isoper(H_td) == true
+        @test QobjEvo(a, coef1) == QobjEvo((a, coef1))
 
         # SuperOperator
         X = a * a'
@@ -205,7 +222,11 @@
         @test isconstant(L_td) == false
         @test issuper(L_td) == true
 
+        coef_wrong1(t) = nothing
+        coef_wrong2(p, t::ComplexF64) = nothing
         @test_logs (:warn,) (:warn,) liouvillian(H_td * H_td) # warnings from lazy tensor
+        @test_throws ArgumentError QobjEvo(a, coef_wrong1)
+        @test_throws ArgumentError QobjEvo(a, coef_wrong2)
         @test_throws MethodError QobjEvo([[a, coef1], a' * a, [a', coef2]])
         @test_throws ArgumentError H_td(ρvec, p, t)
         @test_throws ArgumentError cache_operator(H_td, ρvec)