+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/qutip/QuantumToolbox.jl/tree/main)
 
-
+- Improve the construction of `QobjEvo` ([#338], [#339]):
+    - add new methods to generate `QobjEvo` with input type `Tuple{QuantumObject,Function}`.
+    - add argument checks for each coefficient `Function`.
 
 ## [v0.23.1]
 Release date: 2024-12-06
@@ -53,3 +55,5 @@ Release date: 2024-11-13
 [#324]: https://github.com/qutip/QuantumToolbox.jl/issues/324
 [#330]: https://github.com/qutip/QuantumToolbox.jl/issues/330
 [#335]: https://github.com/qutip/QuantumToolbox.jl/issues/335
+[#338]: https://github.com/qutip/QuantumToolbox.jl/issues/338
+[#339]: https://github.com/qutip/QuantumToolbox.jl/issues/339