+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased](https://github.com/qutip/QuantumToolbox.jl/tree/main)
 
 - Improve the construction of `QobjEvo`. ([#338], [#339])
+- Support `Base.zero` and `Base.one` for `AbstractQuantumObject`. ([#342], [#346])
 
 ## [v0.23.1]
 Release date: 2024-12-06