+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased](https://github.com/qutip/QuantumToolbox.jl/tree/main)
 
 - Improve the construction of `QobjEvo`. ([#338], [#339])
+- Support `Base.zero` and `Base.one` for `AbstractQuantumObject`. ([#342], [#346])
 
 ## [v0.23.1]
 Release date: 2024-12-06
@@ -55,3 +56,5 @@ Release date: 2024-11-13
 [#335]: https://github.com/qutip/QuantumToolbox.jl/issues/335
 [#338]: https://github.com/qutip/QuantumToolbox.jl/issues/338
 [#339]: https://github.com/qutip/QuantumToolbox.jl/issues/339
+[#342]: https://github.com/qutip/QuantumToolbox.jl/issues/342
+[#346]: https://github.com/qutip/QuantumToolbox.jl/issues/346