+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/qutip/QuantumToolbox.jl/tree/main)
 
+
+## [v0.24.0]
+Release date: 2024-12-13
+
 - Improve the construction of `QobjEvo`. ([#338], [#339])
 - Support `Base.zero` and `Base.one` for `AbstractQuantumObject`. ([#342], [#346])
 - Introduce visualization and function `plot_wigner` for easy plotting of Wigner functions. ([#86], [#292], [#347])